refactor(script): split quest buildins into buildin_quest.c
Issue #61 ninth step: move quest helpers out of script.c.
