+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ----------------------------------------
+//1612 [2026/08/09] by Cocoa
+
+・script.c のビルドイン命令分割（第9弾）: quest 系を buildin_quest.c へ分離（挙動変更なし）（script.c, buildin_quest.c, map-server.vcxproj）
+----------------------------------------
 //1611 [2026/08/09] by Cocoa
 
 ・script.c のビルドイン命令分割（第8弾）: CSV 系命令を buildin_csv.c へ分離。csvinit/final・getarraysize を script_internal.h 経由で共有。csvreadarray/csvwritearray の str_buf 直参照を get_str() に置換（挙動変更なし）（script.c, buildin_csv.c, script_internal.h, map-server.vcxproj）
