+++ vcproj/map-server.vcxproj.filters
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
     <ClCompile Include="..\src\map\sql\mapreg_sql.c">
@@ -128,6 +128,9 @@
       <Filter>map_sql</Filter>
     </ClCompile>
     <ClCompile Include="..\src\map\buildin_csv.c">
+      <Filter>map_sql</Filter>
+    </ClCompile>
+    <ClCompile Include="..\src\map\buildin_quest.c">
       <Filter>map_sql</Filter>
     </ClCompile>
     <ClCompile Include="..\src\map\skill.c">
